Fix carrier on groups
* When a new picking was created in the "super" call, it would get the
carrier of the sale order (default implementation of
`_get_new_picking_values` in delivery).
* Modify the way to upgrade the groups and carriers during release:

  * filter moves without available quantity or which do not need a
  release, it prevents assigning a carrier on a picking with nothing
  available or for which we already set the carrier
  * as soon as a picking is assigned a new carrier, push it in a new
  group
  * the new group has the delivery method in the name, to prevent
  confusion with identical group names (default is the SO name)
  * after "super", new picking could have been created for the
  remaining, search them and assign them back to the old group and
  carrier
