+++ delivery_carrier_preference/models/__init__.py
@@ -1,4 +1,5 @@
 from . import delivery_carrier_preference
+from . import procurement_group
 from . import stock_move
 from . import stock_picking
 from . import stock_location_route
